Endless Runner 3D game with bonus Balloon Game(sounds animations)
In this game User is running and obstacles will randomly spawn on their way. User can jump(single and double jump) to avoid the obstacles. User can also increase speed by pressing Left Shift key while running. the score is also increasing with the time. Extra game of balloon in which user has to collect coins and avoid bomb in order to score. In both, once the player bump into an obstacle, user will die and game will be over. Game comes with different animations( running, jumping, death) and sound as well. Download and import the package to enjoy building and learning.